refactor image recognition
