+++ mas_perception_libs/ros/scripts/image_recognition_server
@@ -2,37 +2,7 @@
 import os
 from importlib import import_module
 import rospy
-from mcr_perception_msgs.srv import ImageRecognition, ImageRecognitionResponse
-from mas_perception_libs import ImageClassifier
-
-
-class RecognizeImageService(object):
-    def __init__(self, classifer_class, model_dir, service_name):
-        if not issubclass(classifer_class, ImageClassifier):
-            raise ValueError('classifer is not of ImageClassifier type')
-
-        self._classifier_class = classifer_class
-        self._model_dir = model_dir
-
-        # dictionary of classifiers - key is the model_path
-        self._classifiers = {}
-        self._recog_service = rospy.Service(service_name, ImageRecognition, self.handle_recognize_image)
-
-    def handle_recognize_image(self, req):
-        print('\nnumber of images: ' + str(len(req.images)) + '\n')
-        if req.model_name not in self._classifiers:
-            model_path = os.path.join(self._model_dir, req.model_name + '.h5')
-            class_file = os.path.join(self._model_dir, req.model_name + '.txt')
-            rospy.loginfo('recognition model path: ' + model_path)
-            rospy.loginfo('recognition class file path: ' + class_file)
-            self._classifiers[req.model_name] = self._classifier_class(model_path=model_path, class_file=class_file)
-
-        indices, classes, probabilities = self._classifiers[req.model_name].classify(req.images)
-        response = ImageRecognitionResponse()
-        response.indices = indices
-        response.classes = classes
-        response.probabilities = probabilities
-        return response
+from mas_perception_libs import RecognizeImageService
 
 
 if __name__ == '__main__':
