+++ mas_perception_libs/ros/src/mas_perception_libs/__init__.py
@@ -1,8 +1,7 @@
 from bounding_box import BoundingBox, BoundingBox2D
 from constants import Constant
-from keras_image_classifier import KerasImageClassifier
-from image_classifier import ImageClassifier, ImageClassifierTest
-from image_recognition_service_proxy import ImageRecognitionServiceProxy
+from image_classifier import ImageClassifier, ImageClassifierTest, KerasImageClassifier
+from image_recognition_service import ImageRecognitionServiceProxy, RecognizeImageService
 from image_detector import ImageDetectionKey, ImageDetector, ImageDetectorTest, ImageDetectionService
 from object_detector import ObjectDetector
 from scene_detection_action import SceneDetectionActionServerTest, SceneDetectionActionServer
